refactor: move monadic infrastructure to its own module
